[Refactor] change security componentScan range and permission endpoint
