+++ member-api/src/main/kotlin/io/waterkite94/stalk/api/security/config/UserService.kt
@@ -1,7 +1,8 @@
-package io.waterkite94.stalk.api.security.util
+package io.waterkite94.stalk.api.security.config
 
 import io.waterkite94.stalk.domain.model.Member
 import io.waterkite94.stalk.usecase.usecase.UserSecurity
+import org.springframework.context.annotation.ComponentScan
 import org.springframework.security.core.userdetails.User
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
@@ -9,6 +10,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException
 import org.springframework.stereotype.Component
 
 @Component
+@ComponentScan(basePackageClasses = [UserSecurity::class])
 class UserService(
     private val userSecurity: UserSecurity
 ) : UserDetailsService {
