+++ member-common/encrypt/src/main/kotlin/io/waterkite94/stalk/encrypt/config/PasswordEncoderConfig.kt
@@ -1,12 +1,12 @@
 package io.waterkite94.stalk.encrypt.config
 
 import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.security.crypto.password.PasswordEncoder
-import org.springframework.stereotype.Component
 
-@Component
-class PasswordEncoderConfig {
+@Configuration
+internal class PasswordEncoderConfig {
     @Bean
     fun passwordEncoder(): PasswordEncoder = BCryptPasswordEncoder()
 }
